# docs/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "CDNFoundry Control Plane API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
        "/admin/audit-logs": {
            "get": {
                "operationId": "audit.log.controller",
                "tags": [
                    "Administrator"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationError"
                    }
                }
            }
        },
...
        "/admin/dns/clusters/{cluster}/disable": {
            "post": {
                "operationId": "dns.cluster.controller.disable",
                "tags": [
                    "Administrator"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationError"
                    }
                }
            }
        },
...
        "/admin/domains/{domain}/force-verify": {
            "post": {
                "operationId": "domain.verification.controller",
                "tags": [
                    "Administrator"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationError"
                    }
                }
            }
        },
...
        "/admin/domains/{domain}/users/{user}": {
            "delete": {
                "operationId": "domain.user.controller.destroy",
                "tags": [
                    "Administrator"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationError"
                    }
                }
            }
        },
...
        "/admin/users/{user}": {
            "delete": {
                "operationId": "user.controller.destroy",
                "tags": [
                    "Administrator"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationError"
                    }
                }
            },
            "get": {
                "operationId": "user.controller.show",
                "tags": [
                    "Administrator"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationError"
                    }
                }
            },
            "patch": {
                "operationId": "user.controller.update",
                "tags": [
                    "Administrator"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationError"
                    }
                }
            }
        },
        "/admin/users/{user}/disable": {
            "post": {
                "operationId": "user.controller.disable",
                "tags": [
                    "Administrator"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationError"
                    }
                }
            }
        },
...
        "/domains/{domain}": {
            "delete": {
                "operationId": "domain.controller.destroy",
                "tags": [
                    "Account"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationError"
                    }
                }
            },
            "get": {
                "operationId": "domain.controller.show",
                "tags": [
                    "Account"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationError"
                    }
                }
            },
            "patch": {
                "operationId": "domain.controller.update",
                "tags": [
                    "Account"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationError"
                    }
                }
            }
        },
...
        "/domains/{domain}/disable": {
            "post": {
                "operationId": "domain.controller.disable",
                "tags": [
                    "Account"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationError"
                    }
                }
            }
        },
...
        "/domains/{domain}/dns/records/{record}": {
            "delete": {
                "operationId": "dns.record.controller.destroy",
                "tags": [
                    "Account"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationError"
                    }
                }
            },
            "get": {
                "operationId": "dns.record.controller.show",
                "tags": [
                    "Account"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationError"
                    }
                }
            },
            "patch": {
                "operationId": "dns.record.controller.update",
                "tags": [
                    "Account"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationError"
                    }
                }
            }
        },
...
        "/me/tokens/{token}": {
            "delete": {
                "operationId": "token.controller.destroy",
                "tags": [
                    "Account"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationError"
                    }
                }
            }
        },
        "/nameservers": {
            "get": {
                "operationId": "nameserver.controller",
                "tags": [
                    "Account"
                ],
                "responses": {
                    "200": {
                        "description": "Successful response"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationError"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        },
        "responses": {
            "ValidationError": {
                "description": "The provided data is invalid.",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/ValidationError"
                        }
                    }
                }
            }
        },
        "schemas": {
            "ValidationError": {
                "type": "object",
                "required": [
                    "message",
                    "code",
                    "errors"
                ],
                "properties": {
                    "message": {
                        "type": "string"
                    },
                    "code": {
                        "type": "string",
                        "const": "validation_failed"
                    },
                    "errors": {
                        "type": "object",
                        "additionalProperties": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    }
}